refactor: adhere to max phpstan

+++ src/Indexer.php
@@ -2,6 +2,7 @@
 
 namespace Swis\Laravel\Fulltext;
 
+use Illuminate\Database\Eloquent\Model;
 use Swis\Laravel\Fulltext\Contracts\Indexable;
 
 class Indexer
@@ -13,8 +14,13 @@ class Indexer
 
     public function unIndexOneByClass(string $class, string|int $id): void
     {
+        if (!is_subclass_of($class, Model::class)) {
+            return;
+        }
+
         $record = IndexedRecord::query()
-            ->where('indexable_id', $id)->where('indexable_type', (new $class)->getMorphClass());
+            ->where('indexable_id', $id)
+            ->where('indexable_type', (new $class)->getMorphClass());
 
         if ($record->exists()) {
             $record->delete();
@@ -23,20 +29,27 @@ class Indexer
 
     public function indexOneByClass(string $class, string|int $id): void
     {
+        if (!is_subclass_of($class, Model::class)) {
+            return;
+        }
+
         $model = call_user_func([$class, 'find'], $id);
-        if (in_array(Indexable::class, class_uses($model), true)) {
+        if ($model instanceof Indexable) {
             $this->indexModel($model);
         }
     }
 
     public function indexAllByClass(string $class): void
     {
+        if (!is_subclass_of($class, Model::class)) {
+            return;
+        }
+
         $model = new $class;
-        $self = $this;
-        if (in_array(Indexable::class, class_uses($model), true)) {
-            $model->chunk(100, function ($chunk) use ($self) {
+        if ($model instanceof Indexable) {
+            $model->newQuery()->chunk(100, function ($chunk) {
                 foreach ($chunk as $modelRecord) {
-                    $self->indexModel($modelRecord);
+                    $this->indexModel($modelRecord);
                 }
             });
         }
